Closes ASM-20: Create main TF configuration Closes ASM-22: Create TF manager node module Closes ASM-23: Create TF worker node module
Adds a module for the master manager node which initializes the swarm.
Adds a helper script to fetch the tokens that will be needed by the other manager and worker nodes to join.
Adds a module that is common to both non-master manager nodes and to worker nodes.
Updates security rules so that worker nodes are able to join.